# weather-api/weather-server-spring/src/main/java/io/swagger/api/HistoricalApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.13).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.WeatherHistory;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2019-11-02T10:07:25.240Z[GMT]")
@Api(value = "historical", description = "the historical API")
public interface HistoricalApi {

    @ApiOperation(value = "retrieves weather history", nickname = "getWeatherHistory", notes = "Retrieves a premade dataset of weather history for a specified area ", response = WeatherHistory.class, tags={ "users", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "a premade set of weather history", response = WeatherHistory.class),
        @ApiResponse(code = 400, message = "bad input parameter") })
    @RequestMapping(value = "/historical/{area}/{dataset}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<WeatherHistory> getWeatherHistory(@ApiParam(value = "location of weather data to use",required=true) @PathVariable("area") String area,@ApiParam(value = "dataset number to use",required=true) @PathVariable("dataset") Integer dataset);

}